% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\alias{tietok_dir}
\alias{hakut_dir}
\title{The directory path for Tietokannat}
\usage{
tietok_dir()
}
\description{
The directory path for Tietokannat and Hakutiedot
read from options \code{tietok_dir} and \code{hakut_dir} 
and set them if empty with defaults tietok_dir = "Y:/Ennuste/Taulut KT/Tietokannat"
and hakut_dir = file.path(tietok_dir, "Hakutiedot").
}
\keyword{IO}
