% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTyoelakei.R
\name{getTyoelakei}
\alias{getTyoelakei}
\title{Get tyoelakeindeksi}
\usage{
getTyoelakei(span = NULL)
}
\arguments{
\item{span}{A numeric vector for years to download c(start,end). With NULL, the default, all available years are downloaded. The end year can be omited, and all available years from start are downloaded.}
}
\description{
Get tyoelakeindeksi from \url{http://tyoelakelakipalvelu.etk.fi/fi/indeksi/indeksiluvut.html}
}
\seealso{
Other tyoelakei: \code{\link{makeTyoelakei}}
}
\keyword{IO}
