% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQuery.R
\name{makeQuery}
\alias{makeQuery}
\title{Main function to make Rsuhdanne queries}
\usage{
makeQuery(datab, span = NULL)
}
\arguments{
\item{datab}{the name of the database to make query}

\item{span}{the time interval \code{c(start, end)} for queries if missing, \code{end} can be ommited. 
Will no overwrite existing \code{span}s defined in queries.}
}
\description{
Main function to make Rsuhdanne queries.
}
\keyword{IO}
