% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTyoQ.R
\encoding{UTF-8}
\name{makeTyoQ}
\alias{makeTyoQ}
\title{Read Tyovoimatiedustelun maakuntatauluja}
\usage{
makeTyoQ(vuosi, mk = c("Koko maa", "Uusimaa", "Itä-Uusimaa",
  "Varsinais-Suomi", "Satakunta", "Kanta-Häme", "Pirkanmaa", "Päijät-Häme",
  "Kymenlaakso", "Etelä-Karjala", "Etelä-Savo", "Pohjois-Savo",
  "Pohjois-Karjala", "Keski-Suomi", "Etelä-Pohjanmaa", "Pohjanmaa",
  "Keski-Pohjanmaa", "Pohjois-Pohjanmaa", "Kainuu", "Lappi", "Ahvenanmaa"),
  file1 = file.path(tietok_dir(), "SarjatTK/MK_Työlliset.xlsx"),
  file2 = file.path(tietok_dir(), "SarjatTK/MK_Työlliset_TOL.xlsx"),
  tiedot1 = c("Työikäinen väestö", "Työvoima", "Työlliset",
  "Työttömät"),
  TOL1 = c("A, B  Maatalous, metsätalous, kalatalous; kaivostoiminta (01-09)",
  "C-E  Teollisuus; sähkö-, lämpö-, vesi- ja jätehuolto yms. (10-39)",
  "F  Rakentaminen (41-43)",
  "G  Tukku- ja vähittäiskauppa; moottoriajoneuvojen ja moottoripyörien korjaus (45-47)",
  "H   Kuljetus ja varastointi (49-53)",
  "I   Majoitus- ja ravitsemistoiminta (55-56)",
  "J   Informaatio ja viestintä (58-63)",
  "K, L  Rahoitus- ja vakuutustoiminta; kiinteistöala (64-68)",
  "M, N  Liike-elämän palvelut (69-82)",
  "O   Julkinen hallinto ja maanpuolustus; pakollinen sosiaalivakuutus (84)",   
    "P   Koulutus (85)", "Q   Terveys- ja sosiaalipalvelut (86-88)",
  "R-U  Muu palvelutoiminta (90-99)", "Toimialat yhteensä (00-99)"),
  TOL2 = c("A, B   Maa-, metsä- ja kalatalous; kaivostoiminta (01-09)",
  "C-E   Teollisuus; sähkö-, lämpö-, vesi- ja jätehuolto yms. (10-39)",
  "F   Rakentaminen (41-43)",
  "G   Tukku- ja vähittäiskauppa; moottoriajoneuvojen ja moottoripyörien korjaus (45-47)",
  "H   Kuljetus ja varastointi (49-53)",
  "I   Majoitus- ja ravitsemistoiminta (55-56)",
  "J   Informaatio ja viestintä (58-63)",
  "K, L   Rahoitus- ja vakuutustoiminta; kiinteistöala (64-68)",
  "M, N   Liike-elämän palvelut (69-82)",
  "O   Julkinen hallinto ja maanpuolustus; pakollinen sosiaalivakuutus (84)",   
    "P   Koulutus (85)", "Q   Terveys- ja sosiaalipalvelut (86-88)",
  "R-U   Muu palvelutoiminta (90-99)", "Toimialat yhteensä (00-99)"),
  TOL3 = c("A, B   Maatalous, metsätalous, kalatalous; kaivostoiminta (01-09)",
  "C-E   Teollisuus; sähkö-, lämpö-, vesi- ja jätehuolto yms. (10-39)",
  "F   Rakentaminen (41-43)",
  "G   Tukku- ja vähittäiskauppa; moottoriajoneuvojen ja moottoripyörien korjaus (45-47)",
  "H   Kuljetus ja varastointi (49-53)",
  "I   Majoitus- ja ravitsemistoiminta (55-56)",
  "J   Informaatio ja viestintä (58-63)",
  "K, L   Rahoitus- ja vakuutustoiminta; kiinteistöala (64-68)",
  "M, N   Liike-elämän palvelut (69-82)",
  "O   Julkinen hallinto ja maanpuolustus; pakollinen sosiaalivakuutus (84)",   
    "P   Koulutus (85)", "Q   Terveys- ja sosiaalipalvelut (86-88)",
  "R-U   Muu palvelutoiminta (90-99)", "Toimialat yhteensä (00-99)"),
  TOL4 = c("A, B   Maatalous, metsätalous, kalatalous; kaivostoiminta (01-09)",
  "C-E   Teollisuus; sähkö-, lämpö-, vesi- ja jätehuolto yms. (10-39)",
  "F   Rakentaminen (41-43)",
  "G   Tukku- ja vähittäiskauppa; moottoriajoneuvojen ja moottoripyörien korjaus (45-47)",
  "H   Kuljetus ja varastointi (49-53)",
  "I   Majoitus- ja ravitsemistoiminta (55-56)",
  "J   Informaatio ja viestintä (58-63)",
  "K, L   Rahoitus- ja vakuutustoiminta; kiinteistöala (64-68)",
  "M, N   Liike-elämän palvelut (69-82)",
  "O   Julkinen hallinto ja maanpuolustus; pakollinen sosiaalivakuutus (84)",   
    "P   Koulutus (85)", "Q   Terveys- ja sosiaalipalvelut (86-88)",
  "R-U   Taiteet, viihde ja virkistys; muu palvelutoiminta (90-99)",
  "Toimialat yhteensä (00-99)"))
}
\arguments{
\item{vuosi}{a numeric, a year for time-series to start}

\item{mk}{A vector for maakunnat}

\item{file1}{a string, filepath for the first query to write}

\item{file2}{a string, filepath for the second query to write}

\item{tiedot1}{a vector for variables in the first query}

\item{TOL1}{a vector for the first set of TOL names in the second query}

\item{TOL2}{a vector for the second set of TOL names in the second query}

\item{TOL3}{a vector for the 3. set of TOL names}

\item{TOL4}{a vector for the 4 set of TOL names}
}
\description{
read files based on their vars list. Add Itä-Uusimaa to Uusimaa if prensent 
Set a file specific vars list because variables and they names change. If
changes in future code on this function might have to be changed. Calls
\code{\link{read_tyoQ}}
}
\seealso{
Other tyoQ: \code{\link{make_TyoQ_1_rds}},
  \code{\link{read_tyoQ}}, \code{\link{tyoQfiles}}
}
\keyword{IO}
