% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_px.R
\name{read_px}
\alias{read_px}
\title{Read PC-Axis files}
\usage{
read_px(file, in.codes = FALSE, org.order = FALSE, encoding = "unknown")
}
\arguments{
\item{file}{a file or connection to read. See \code{\link{scan}}}

\item{in.codes}{A locigal, if \code{TRUE} also CODES are included in
addition to VALUES to identify data.}

\item{org.order}{A locigal, if \code{TRUE} VALUES and CODES factors are
given in original order. Otherwise ordering based on normal
\code{\link{factor}} manner.}

\item{encoding}{a encoding for the \code{\link{scan}}.}
}
\value{
\code{data.frame}. TODO: In future with metadata attributes.
}
\description{
The funtion to read PC-Axis px-files.
}
\examples{
  \dontrun{
    x <- read_px("http://pxweb2.stat.fi/database/StatFin/tul/tvt/2009/010_tvt_2009_2011-02-18_tau_101_fi.px", org.order = TRUE)
  }
}
\references{
\url{http://www.scb.se/Pages/StandardNoLeftMeny____314045.aspx}
}
\author{
Janne Huovari
}
\keyword{IO}
\keyword{database}
