% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tyoQ.R
\name{read_tyoQ}
\alias{read_tyoQ}
\title{Read Tyovoimatiedustelun maakuntatauluja}
\usage{
read_tyoQ(files, vars, year)
}
\arguments{
\item{files}{A vector of files to read}

\item{vars}{a list of vars lists for \code{\link{dload.tk}}}

\item{year}{a numeric, a year for time-series to start}
}
\description{
read files based on their vars list. Add Itä-Uusimaa to Uusimaa if prensent
}
\seealso{
Other tyoQ: \code{\link{makeTyoQ}},
  \code{\link{make_TyoQ_1_rds}}, \code{\link{tyoQfiles}}
}
\keyword{IO}
