% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyoQfiles.R
\name{tyoQfiles}
\alias{tyoQfiles}
\title{Get files for read_tyoQ}
\usage{
tyoQfiles(pattern,
  ddir = "Y:/Yhteiset datat/Aluetalous/Työ/Työvoimatutkimus MK")
}
\arguments{
\item{pattern}{pattern in files to match}

\item{ddir}{folder to look for files.}
}
\description{
Get data files for \code{\link{make_TyoQ_1_rds}} from \code{ddir}
}
\examples{
\dontrun{
 files1 <- tyoQfiles("Maakunta1")
 files2 <- tyoQfiles("Maakunta2")
 }
}
\seealso{
Other tyoQ: \code{\link{make_TyoQ_1_rds}}
}
\keyword{IO}
