% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_shadow_rate.R
\name{update_shadow_rate}
\alias{update_shadow_rate}
\title{Update shadow policy rates}
\usage{
update_shadow_rate()
}
\description{
download files from http://faculty.chicagobooth.edu/jing.wu/research/data/WX.html
and copy to "Y:/Ennuste/Taulut KT/Tietokannat/SarjatKäsin/Wu & Xia (2013) shadow rates"
Macrobond in-house links there
}
