% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateHaut.R
\name{updateHaut}
\alias{updateHaut}
\title{Write updated haku- file for ETLA}
\usage{
updateHaut(file, freq, final = FALSE)
}
\arguments{
\item{file}{haut file name to update}

\item{freq}{a character. frequency of haut, see \code{\link{dload.etla}}}

\item{final}{a locigal. \code{FALSE} (default) print new ones separetly below old ones with holes in data. 
Easier to identify new ones. \code{TRUE} makes usable haku-file.}
}
\description{
Write updated haku-file in \code{\link{hakut_dir}} directory.
Read old one and write new one with old and new serie names by name \code{oldname_uusi}.
}
\note{
At the moment \code{rbind} old and new ones, and remove dublicates. 
Update have to be finished manually.
}
\examples{
\dontrun{
 updateHaut("KT Y.csv", "Y")}
}
\keyword{IO}
