% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeMQuery.R
\name{writeMQuery}
\alias{writeMQuery}
\title{Write multiple a query lists}
\usage{
writeMQuery(qnames, query, group, datab, multiVars)
}
\arguments{
\item{qnames}{a character vector of queries to write.}

\item{query}{a list. The query to write. \code{\link{makeQuery}} passes to \code{\link{dload}}. Could
include item \code{trans} that is list of functions that perform transformation to series 
before saving to file. the \code{trans} is not passed to \code{\link{dload}}.}

\item{group}{the name of the query group. Also name of the query-file and writen xlsx-file}

\item{datab}{the name of the database. \code{\link{makeQuery}} passes to \code{\link{dload}}}

\item{multiVars}{a character. The name of the \code{vars} list item which will be writen to multiple queries.
The list item value should correspond \code{qnames}.}
}
\value{
Invisibly \code{TRUE}
}
\description{
Help to write multiple similar query lists with \code{writeQuery}. See \code{writeQuery} for details.
}
\examples{
  qu <- list(table = "namq_gdp_k",
               vars = list(
                   s_adj=c("SWDA"),
                   unit=c("PCH_PRE"),
                   geo=c("DE","EA","FI","FR","SE","UK","US","JP"),
                   indic_na = c("b1gm", "p31_s14_s15")
               )
             )
  writeMQuery(qnames = c("bkt", "y kulutus"), query = qu, group = "koe", datab = "eurostat", multiVars = "indic_na")
  koe <- readQuery("koe", "eurostat")
  unlink(file.path(hakut_dir(), "eurostat", "koe")) #delete test-file
}
\keyword{IO}
