% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQuery.R, R/writeQuery.R
\name{readQuery}
\alias{readQuery}
\alias{writeQuery}
\title{Write and read a query list}
\usage{
readQuery(group, datab)

writeQuery(qname, query, group, datab)
}
\arguments{
\item{group}{the name of the query group. Also name of the query-file and writen xlsx-file}

\item{datab}{the name of the database. \code{\link{makeQuery}} passes to \code{\link{dload}}}

\item{qname}{The name of the query to write.}

\item{query}{a list. The query to write. \code{\link{makeQuery}} passes to \code{\link{dload}}. Could
include item \code{trans} that is list of functions that perform transformation to series 
before saving to file. the \code{trans} is not passed to \code{\link{dload}}.}
}
\value{
\code{readQuery} return list with saved queries.
}
\description{
\code{writeQuery} writes a query list to the file in Hakutiedot directory. 
\code{readQuery} reads a queries list from the file.
}
\details{
Write query lists to the file. Append existing queries. Queries can be deleted with
\code{query = NULL}. Uses \code{\link{dput}} and \code{\link{dget}}.
}
\examples{
  qu <- list(table = "namq_gdp_k",
               vars = list(
                   s_adj=c("SWDA"),
                   unit=c("PCH_PRE"),
                   geo=c("DE","EA","FI","FR","SE","UK","US","JP"),
                   indic_na=c("B1GM")
               )
             )
  writeQuery("na q", qu, "koe", "eurostat")
  writeQuery("na q2", qu, "koe", "eurostat")
  writeQuery("na q2", NULL, "koe", "eurostat")
  koe <- readQuery("koe", "eurostat")
  unlink(file.path(hakut_dir, "eurostat", "koe")) #delete test-file
}
\keyword{IO}
